realtime_service/requirements.txt
fastapi==0.115.12
redis-async==1.4.0
uvicorn==0.34.3
